mask_from_shape: cube coordinate system optional
Some cubes does not have a coordinate system, e.g. if they are not on a
rotated pole projection. For instance the S-Lens ensemble. For the cubes
we don't need to convert the coordinate to create the mask from the
shapefile since they are already in the same format. This conversion can
now be turned off with the `coord_system` arg.